+++ @remirror/core-extensions/src/extensions/ssr-helpers/index.ts
@@ -1,2 +1,2 @@
-export * from './utils';
-export * from './ssr-helpers.extension';
+export * from './ssr-helpers-utils';
+export * from './ssr-helpers-extension';