+++ @remirror/core/src/nodes/text-extension.ts
@@ -1,5 +1,10 @@
 import { NodeExtension } from '../node-extension';
 
+/**
+ * The default text passed into the prosemirror schema.
+ *
+ * @builtin
+ */
 export class TextExtension extends NodeExtension {
   get name() {
     return 'text' as const;